+++ linkerscope.py
@@ -1,11 +1,17 @@
 #!/usr/bin/env python3
 
 import argparse
+import copy
 import yaml
+
+from area_view import AreaView
+from helpers import safe_element_get
+from labels import Labels
+from links import Links
 from map_drawer import Map
 from style import Style
 from map_file_parser import MapFileParser
-from sectionsview import SectionsView, Sections
+from sections import Sections
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--output',
@@ -23,9 +29,6 @@ parser.add_argument('--configuration',
 
 args = parser.parse_args()
 
-sections = MapFileParser(args.input).parse()
-
-config = []
 areas = []
 
 with open(args.configuration, 'r') as file:
@@ -35,36 +38,34 @@ if config['areas'] is None:
     print('No information to show on current configuration file')
     exit(-1)
 
-for element in config['areas']:
-    area = element.get('area')
-
-    filtered_sections = (Sections(sections=sections)
-     .filter_address_min(area.get('address', {}).get('min'))
-     .filter_address_max(area.get('address', {}).get('max'))
-     .filter_size_min(area.get('size', {}).get('min'))
-     .filter_size_max(area.get('size', {}).get('max'))
-     )
-
-    if len(filtered_sections.sections) == 0:
-        print("Filtered sections produced no results")
-        continue
-
-    sections_view = SectionsView(
-        sections=filtered_sections.get_sections(),
-         area=area)
-
-    if len(sections_view.sections) == 0:
-        print("Current view doesn't show any section")
-        continue
-    areas.append(sections_view)
+    # TODO: linked sections compatibility
 
 base_style = Style().get_default()
-base_style.override_properties_from(Style(style=config.get('style')))
+base_style.override_properties_from(Style(style=config.get('style', None)))
 
-a = Map(diagrams=areas,
-        links=config.get('links'),
-        style=base_style
-        )
+for element in config['areas']:
+    area_style = copy.deepcopy(base_style)
 
-a.draw(args.output)
+    area = element.get('area')
+    section_size = area.get('section-size', {})
+    _range = area.get('range', {})
+    area_view = AreaView(
+        sections=(Sections(sections=MapFileParser(args.input).parse())
+                  .filter_address_min(safe_element_get(_range, 0))
+                  .filter_address_max(safe_element_get(_range, 1))
+                  .filter_size_min(safe_element_get(section_size, 0))
+                  .filter_size_max(safe_element_get(section_size, 1))
+                  ),
+        area_config=area,
+        global_config=config,
+        style=area_style.override_properties_from(Style(style=area.get('style')))
+    )
+    areas.extend(area_view.get_processed_section_views())
+yaml_links = config.get('links', None)
 
+links_style = copy.deepcopy(base_style)
+links = Links(config.get('links', {}), style=links_style.override_properties_from(Style(style=yaml_links.get('style') if yaml_links is not None else None)))
+Map(area_view=areas,
+    links=links,
+    style=base_style,
+    file=args.output).draw()