+++ map_file_parser.py
@@ -1,6 +1,6 @@
 import yaml
 
-from area import Section
+from section import Section
 from map_parser import MapParser
 import os
 
@@ -21,7 +21,6 @@ class MapFileParser:
     def parse_yaml(filename):
         sections = []
 
-        address_max = 0x60000
         with open(filename, 'r') as file:
             y = yaml.safe_load(file)