+++ style.py
@@ -1,20 +1,31 @@
 class Style:
-    section_fill_color: str
-    section_stroke_color: str
-    section_stroke_width: int
-    link_stroke_width: int
-    link_stroke_color: str
-    label_font: str
-    label_color: int
-    label_size: int
-    label_stroke_width: int
-    map_background_color: int
-    background_color: int
+    # Non SVG
+    background: str
+    break_type: str
+    break_size: int
+    growth_arrow_size: float
+    growth_arrow_stroke: str
+    growth_arrow_fill: str
+    stroke_dasharray: str
+    # SVG
+    fill: str
+    stroke: str
+    stroke_width: int
+    size: int
+    font_size: int
+    font_type: str
+    weight: int
+    opacity: int
 
+    text_stroke: str
+    text_stroke_width: int
+    text_fill: str
+
+    weigth: int
     def __init__(self, style=None):
         if style is not None:
             for key, value in style.items():
-                setattr(self, key, style.get(key, value))
+                setattr(self, key.replace('-','_'), style.get(key, value))
 
     def override_properties_from(self, style):
         '''
@@ -31,17 +42,36 @@ class Style:
             value = getattr(style, member)
             setattr(self, member, value)
 
+        return self
+
     @staticmethod
     def get_default():
         default_style = Style()
-        default_style.section_fill_color = '#CCE5FF'
-        default_style.section_stroke_color = '#3399FF'
-        default_style.section_stroke_width = 2
-        default_style.label_color = 'blue'
-        default_style.label_size = '16px'
-        default_style.label_stroke_width = 1
-        default_style.link_stroke_width = 1
-        default_style.link_stroke_color = 'grey'
-        default_style.map_background_color = '#CCCCFF'
-        default_style.background_color = 'white'
+        default_style.break_type = '≈'
+        default_style.break_size = 20
+
+        default_style.growth_arrow_size = 1
+
+        default_style.background = 'white'
+        default_style.fill = '#CCE5FF'
+        default_style.stroke = 'black'
+        default_style.stroke_width = 1
+        default_style.size = 2
+
+        default_style.font_size = 16
+        default_style.font_type = 'Helvetica'
+
+        default_style.weight = 1
+        default_style.opacity = 1
+
+        default_style.text_stroke = 'black'
+        default_style.text_fill = 'black'
+        default_style.text_stroke_width = 0
+
+        default_style.fill = 'black'
+        default_style.growth_arrow_fill = 'white'
+        default_style.growth_arrow_stroke = 'black'
+        default_style.stroke_dasharray = '3,2'
+        default_style.weigth = 2
+
         return default_style